# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "记忆 API",
        "title": "Remembrance API",
        "contact": {
            "name": "lzd",
            "email": "2804366305@qq.com"
        },
        "version": "1.0"
    },
    "host": "8.138.81.141",
    "basePath": "api",
    "paths": {
...
        "/api/user/changename": {
            "post": {
                "description": "保证前后两次邮箱相同，并将邮箱与更改后的昵称上传",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "更改昵称",
                "parameters": [
                    {
                        "description": "userid",
                        "name": "userid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_User"
                        }
                    },
                    {
                        "description": "name",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/user/changepassword": {
            "post": {
                "description": "检查验证码正确后，将 邮箱 与更改后的 密码 上传 （保证检查验证码时的邮箱，与上传的邮箱一致）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "更改密码",
                "parameters": [
                    {
                        "description": "email",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_User"
                        }
                    },
                    {
                        "description": "password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/user/getinfo": {
            "get": {
                "description": "根据userid获取用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "description": "email",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_User"
                        }
                    },
                    {
                        "description": "password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/user/group/creat": {
            "put": {
                "description": "需要创建者的 id 群名 code （目前群名不能重复）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "创建群",
                "parameters": [
                    {
                        "description": "userid",
                        "name": "userid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_Group"
                        }
                    },
                    {
                        "description": "name",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_Group"
                        }
                    },
                    {
                        "description": "code",
                        "name": "code",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_Group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/user/group/delete": {
            "post": {
                "description": "传入被解散的群的id就行，确保只有群主能执行这个操作",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "解散群",
                "parameters": [
                    {
                        "description": "groupid",
                        "name": "groupid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_Group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/user/group/get": {
            "get": {
                "description": "传入userid 获得该用户所加入的群信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取群",
                "parameters": [
                    {
                        "description": "userid",
                        "name": "userid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_Group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/user/group/join": {
            "post": {
                "description": "需要 加入者的id 加入的群名 对应的code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "加入群",
                "parameters": [
                    {
                        "description": "userid",
                        "name": "userid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_Group"
                        }
                    },
                    {
                        "description": "name",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_Group"
                        }
                    },
                    {
                        "description": "code",
                        "name": "code",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_Group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        },
        "/api/user/group/out": {
            "post": {
                "description": "主动退出则传退出者的userid，被踢则传被踢的人的userid  还需要群名",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "退出群",
                "parameters": [
                    {
                        "description": "userid",
                        "name": "userid",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_Group"
                        }
                    },
                    {
                        "description": "name",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_Group"
                        }
                    },
                    {
                        "description": "code",
                        "name": "code",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.S_Group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OkMesData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.FailMesData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.S_Group": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "groupid": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "DeletedAt DeletedAt `gorm:\"index\"`",
                    "type": "string"
                },
                "peoplenum": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.S_User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "pospersonaltnum": {
                    "type": "integer"
                },
                "postcommonnum": {
                    "type": "integer"
                },
                "postgroupnum": {
                    "type": "integer"
                },
                "stampnum": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "description": "DeletedAt DeletedAt `gorm:\"index\"`",
                    "type": "string"
                }
            }
        },
        "response.FailMesData": {
            "type": "object"
        },
        "response.OkMesData": {
            "type": "object"
        }
    }
}